LayoutTests/imported/w3c/web-platform-tests/fetch/stale-while-revalidate/frame-removal.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test that frame removal cancels its revalidations</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<body>
<script>
async_test(function(t) {
  var request_token = token();
  const url = `resources/stale-frame.py?token=` + request_token

  // first frame load
  let frame = document.createElement("iframe");
  frame.src = url;
  frame.onload = () => {
    // second frame load, triggers revalidation
    let frame2 = document.createElement("iframe");
    frame2.src = url;
    frame2.onload = () => {
      // remove frame, canceling revalidation
      frame2.remove();
      // third frame load, triggers new revalidation
      let frame3 = document.createElement("iframe");
      frame3.src = url;
      frame3.onload = () => {
        fetch("resources/stale-frame.py?query&token=" + request_token).then(t.step_func((response) => {
          assert_equals(response.headers.get("Count"), '2');
          t.done();
        }));
      }
      document.body.appendChild(frame3);
    }
    document.body.appendChild(frame2);
  }
  document.body.appendChild(frame);
}, 'Frame removal should cancel its revalidations');
</script>
</body>
